Blinking Funtional
Added
Blinking Node to PlayerShip
Blinking.cs The script containing the code for blinking
Changed:
Playership.cs
PlayerController.cs
Ship.cs
To be added
Second version of blinking for testing

To be changed:
Controlls out of Blinking and only in PlayerController.

Look into:
Spawn location of the Area2D, when it is displayed it seems to be spawning to far away.
When Qeued free this doesn't appear to be an issue